Added conversational mode.
When conv button is pushed, an empty (except for %'s) ".wiz" file is sent to Axis and conversational mode is started. Clicking the check box on a wizard causes the wizard subroutine call to be sent to Axis and executed. There is no status checking done; so the machine must be homed etc for this to work.